Added props to Attendees.jsx
