+++ src/Components/Attendees.jsx
@@ -11,7 +11,7 @@ export default function Attendees({ showAttendees, attendees, event, updateEvent
       <div key={event.id} className="attendees">
         {attendees.map(attendee => (
         // add attendee prop here where HTML code used to be (HTML placed in Attendee component)
-        <Attendee key={attendee.id} attendee={attendee} updateEventAttendance={updateEventAttendance} />
+        <Attendee key={attendee.id} attendee={attendee} event={event} updateEventAttendance={updateEventAttendance} />
         ))}
       </div>
       ) : null}
